Propose date change for September (#107)
The current September date is the day after GraphQL conf, @martinbonnin
and I thought it best to change this date.

I've chosen 25th September as the 18th is earmarked for the secondary
main working group day.

cc @jonathanrainer you'll want to be aware of this date change as you're
on the agenda